extracted createBackupPolicy, getProtectedResourceName, and protectFileshare logic into its own actions
