+++ pkg/skr/azurerwxvolumebackup/createBackupPolicy.go
@@ -1,0 +1,45 @@
+package azurerwxvolumebackup
+
+import (
+	"context"
+	"fmt"
+	cloudresourcesv1beta1 "github.com/kyma-project/cloud-manager/api/cloud-resources/v1beta1"
+	"github.com/kyma-project/cloud-manager/pkg/composed"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+)
+
+func createBackupPolicy(ctx context.Context, st composed.State) (error, context.Context) {
+
+	state := st.(*State)
+	logger := composed.LoggerFromCtx(ctx)
+	backup := state.ObjAsAzureRwxVolumeBackup()
+
+	// short circuit if protectedResourceName already exist
+	if state.protectedResourceName != "" {
+		return nil, ctx
+	}
+
+	fileShareName := state.fileShareName
+	vaultName := state.vaultName
+	resourceGroupName := state.resourceGroupName
+
+	policyName := fmt.Sprintf("%v-backup-policy", fileShareName) // TODO: add to state
+
+	err := state.client.CreateBackupPolicy(ctx, vaultName, resourceGroupName, policyName)
+	if err != nil {
+		logger.Error(err, "failed to create backup policy")
+		backup.Status.State = cloudresourcesv1beta1.AzureRwxBackupError
+		return composed.PatchStatus(backup).
+			SetExclusiveConditions(metav1.Condition{
+				Type:    cloudresourcesv1beta1.ConditionTypeError,
+				Status:  metav1.ConditionTrue,
+				Reason:  cloudresourcesv1beta1.ConditionReasonError,
+				Message: fmt.Sprintf("Could not create BackupPolicy for backup: %s", err),
+			}).
+			FailedError(composed.StopWithRequeue).
+			Run(ctx, state)
+	}
+
+	// Success; continue on
+	return nil, ctx
+}
