+++ pkg/skr/azurerwxvolumebackup/reconciler.go
@@ -44,6 +44,9 @@ func (r *reconciler) newAction() composed.Action {
 				loadPersistentVolume,
 				createClient,
 				createVault,
+				getProtectedResourceName,
+				createBackupPolicy,
+				protectFileshare,
 				createBackup,
 			), nil,
 		),
